Added x64 configuration to VC++. MinGW / G++ 64-bit version still WIP.

+++ src/io/file_io.cpp
@@ -391,7 +391,7 @@ int path::list_files(std::vector<std::string>& out, const char* directory, const
 	} while (FindNextFileA(hFind, &ffd));
 
 	FindClose(hFind);
-	return out.size();
+	return (int)out.size();
 }
